+++ stable_baselines/acktr/acktr_disc.py
@@ -102,6 +102,9 @@ class ACKTR(ActorCriticRLModel):
         if _init_setup_model:
             self.setup_model()
 
+    def _get_pretrain_placeholders(self):
+        raise NotImplementedError()
+
     def setup_model(self):
         with SetVerbosity(self.verbose):