+++ stable_baselines/ddpg/ddpg.py
@@ -269,6 +269,9 @@ class DDPG(OffPolicyRLModel):
         if _init_setup_model:
             self.setup_model()
 
+    def _get_pretrain_placeholders(self):
+        raise NotImplementedError()
+
     def setup_model(self):
         with SetVerbosity(self.verbose):